PR Fedora branch to test latest changes

+++ scripts/set_up_nvim.sh
@@ -1,7 +1,19 @@
 #!/bin/bash
 
-sudo apt install gcc
-sudo snap install nvim --classic
+if [ -f /etc/os-release ]; then
+    . /etc/os-release
+
+    case "$ID" in
+      fedora)
+          ;;
+      arch)
+          ;;
+      ubuntu|debian)
+          sudo apt install gcc
+          sudo snap install nvim --classic
+          ;;
+    esac
+fi
 
 # -n avoids creating a sub-folder inside the old target folder
 ln -sfn $DOTFILES/config/nvim ~/.config/nvim
